refactor to class structure for todoList
