+++ src/types.ts
@@ -1,0 +1,5 @@
+export type Todo = {
+    id: string;
+    text: string;
+    completed: boolean;
+};
